feat(infra): add PostgreSQL server, database, and firewall rule resources to the Terraform configuration
Add a PostgreSQL server, database, and firewall rule resources to the Terraform configuration to enable the application to store data in a PostgreSQL database. The server is created with the name "lynx-order-tracker-dev-db", located in the same resource group as the other resources, and with a SKU of "B_Gen5_1". The database is created with the same name as the server, and with a charset of "UTF8" and a collation of "English_United States.1252". A random password is generated for the PostgreSQL server's administrator login. A firewall rule is created to allow connections from the IP address "95.175.96.152".